+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 1.3.9
 
 - Improved backup of database
+- Improved backup of media and finance files
 - Fixed checkbox UI issues
 
 ## 1.3.8